refactor: extract AddSymbolDialog from watchlist.tsx (#143)
Move dialog state (symbol, debouncedQuery, showSuggestions, suggestionsRef),
debounce effect, and all dialog JSX into a self-contained component.
WatchlistPage just renders <AddSymbolDialog />.
